requirements.txt
psycopg2-binary
pydantic
pydantic-settings